Refractored encoder tests. Each encoder gets its own folder
