+++ tests/encoders/vlad/test_vlad.py
@@ -49,24 +49,24 @@ def vlad_pca(category_train_images_flat: list[np.ndarray]) -> VLADEncoder:
     return encoder
 
 
-def test_vlad_clustering_model_type() -> None:
+def test_clustering_model_type() -> None:
     """VLAD builds a KMeans clustering model."""
     assert isinstance(VLADEncoder(n_clusters=8).clustering_model, KMeans)
 
 
-def test_vlad_n_clusters_param_kwargs_collision() -> None:
+def test_n_clusters_param_kwargs_collision() -> None:
     """Passing ``n_clusters`` inside ``kmeans_params`` raises ``ValueError``."""
     with pytest.raises(ValueError, match="Pass 'n_clusters' directly"):
         VLADEncoder(kmeans_params={"n_clusters": 8})
 
 
-def test_vlad_wrong_weights_class_raises() -> None:
+def test_wrong_weights_class_raises() -> None:
     """Passing GMM weights to VLAD raises ``ValueError``."""
     with pytest.raises(ValueError, match="only pass an instance of KMeansWeights"):
         VLADEncoder(weights=GMMWeights.OXFORD102_K256_SIFT)
 
 
-def test_vlad_clustering_setter_rejects_non_kmeans() -> None:
+def test_clustering_setter_rejects_non_kmeans() -> None:
     """Assigning a non-KMeans clustering model raises ``ValueError``."""
     encoder = VLADEncoder(n_clusters=8)
     with pytest.raises(
@@ -75,7 +75,7 @@ def test_vlad_clustering_setter_rejects_non_kmeans() -> None:
         encoder.clustering_model = GaussianMixtureModel(8)
 
 
-def test_vlad_encode_shape_no_pca(
+def test_encode_shape_no_pca(
     vlad_no_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """Without PCA, a single image encodes to ``(1, n_clusters * 128)``."""
@@ -83,7 +83,7 @@ def test_vlad_encode_shape_no_pca(
     assert out.shape == (1, VLAD_DIM_NO_PCA)
 
 
-def test_vlad_encode_batch_shape(
+def test_encode_batch_shape(
     vlad_no_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """A batch of three images encodes to ``(3, n_clusters * 128)``."""
@@ -92,7 +92,7 @@ def test_vlad_encode_batch_shape(
     assert vlad_no_pca.encode(batch).shape == (3, VLAD_DIM_NO_PCA)
 
 
-def test_vlad_encode_shape_with_pca(
+def test_encode_shape_with_pca(
     vlad_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """With PCA to 32 dims, a single image encodes to ``(1, n_clusters * 32)``."""
@@ -100,7 +100,7 @@ def test_vlad_encode_shape_with_pca(
     assert out.shape == (1, VLAD_DIM_PCA)
 
 
-def test_vlad_encode_rows_l2_normalized(
+def test_encode_rows_l2_normalized(
     vlad_no_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """Each non-zero cluster row of the encoding has unit L2 norm."""
@@ -110,14 +110,14 @@ def test_vlad_encode_rows_l2_normalized(
     assert non_zero == pytest.approx(np.ones_like(non_zero), rel=1e-3)
 
 
-def test_vlad_encode_single_rgb_image_ok(
+def test_encode_single_rgb_image_ok(
     vlad_no_pca: VLADEncoder, rgb_image: ImageObj
 ) -> None:
     """A bare 3-D RGB array is treated as a single image."""
     assert vlad_no_pca.encode(rgb_image.array).shape == (1, VLAD_DIM_NO_PCA)
 
 
-def test_vlad_encode_single_2d_must_be_wrapped(
+def test_encode_single_2d_must_be_wrapped(
     vlad_no_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """A single 2-D image must be wrapped in a list; a bare 2-D array iterates rows."""
@@ -127,7 +127,7 @@ def test_vlad_encode_single_2d_must_be_wrapped(
         vlad_no_pca.encode(gray)
 
 
-def test_vlad_encode_no_descriptors_raises(
+def test_encode_no_descriptors_raises(
     vlad_no_pca: VLADEncoder, solid_image: ImageObj
 ) -> None:
     """Encoding a featureless image raises a clear ``ValueError``."""
@@ -135,7 +135,7 @@ def test_vlad_encode_no_descriptors_raises(
         vlad_no_pca.encode([solid_image.array])
 
 
-def test_vlad_encode_accepts_tensor(
+def test_encode_accepts_tensor(
     vlad_no_pca: VLADEncoder, checkerboard_image: ImageObj
 ) -> None:
     """A grayscale torch tensor image is accepted and encodes like its array."""
@@ -143,7 +143,7 @@ def test_vlad_encode_accepts_tensor(
     assert vlad_no_pca.encode([tensor]).shape == (1, VLAD_DIM_NO_PCA)
 
 
-def test_vlad_predict_before_learn_raises(checkerboard_image: ImageObj) -> None:
+def test_predict_before_learn_raises(checkerboard_image: ImageObj) -> None:
     """Encoding before learning raises ``NotFittedError`` (KMeans not fitted)."""
     with pytest.raises(NotFittedError):
         VLADEncoder(n_clusters=8).encode([checkerboard_image.array])
